+++ R/aMSE.R
@@ -11,7 +11,7 @@
 #' @importFrom makehtml addplot addtable dirExists filenametopath
 #' @importFrom makehtml getextension htmltable logfilename make_html pathend
 #' @importFrom makehtml pathtype setuphtml write_css write_head
-#' @useDynLib aMSE
+#' @useDynLib aMSE, .registration = TRUE
 NULL
 
 #' @title aMSE functions for Conditioning and Running an Invertebrate MSE